pyproject.toml
[project]
name = "items-exchange"
version = "0.1.0"
description = ""
authors = [
    {name = "Ivan Tkachenko",email = "i.tkachenko@zohomail.eu"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "django-filter (>=25.1,<26.0)",
    "ruff (>=0.11.5,<0.12.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
